# docs/openapi.json
{
  "openapi" : "3.0.1",
  "info" : {
    "license" : {
      "name" : "Apache-2.0",
      "url" : "https://www.apache.org/licenses/LICENSE-2.0"
    },
    "title" : "ATP-TDM"
  },
  "security" : [
    {
      "Bearer Authentication" : [ ]
    }
  ],
...
  "paths" : {
...
    "/api/tdm/cleanup/config" : {
      "post" : {
        "description" : "Saves the cleanup configuration for the table in tableName and for the tables of the same system in the environments listed in environmentsList, and schedules it. queryTimeout must be between 1 and EXTERNAL_QUERY_MAX_TIMEOUT seconds (3600 by default); a value out of range returns HTTP 400.",
        "operationId" : "saveCleanupConfig",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/CleanupSettings"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CleanupSettings"
                }
              }
            },
            "description" : "OK"
          }
        },
        "summary" : "Save cleanup settings",
        "tags" : [
          "data-cleanup-controller"
        ]
      }
    },
    "/api/tdm/cleanup/config/{id}" : {
      "get" : {
        "description" : "Returns the cleanup configuration and the environments of the tables that use it. tableName is not set in the response.",
        "operationId" : "getCleanupConfig",
        "parameters" : [
          {
            "description" : "Cleanup configuration ID.",
            "in" : "path",
            "name" : "id",
            "required" : true,
            "schema" : {
              "type" : "string",
              "format" : "uuid"
            }
          }
        ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CleanupSettings"
                }
              }
            },
            "description" : "OK"
          }
        },
        "summary" : "Get cleanup settings",
        "tags" : [
          "data-cleanup-controller"
        ]
      }
    },
...
    "/api/tdm/cleanup/next/run" : {
      "get" : {
        "description" : "Returns the next time the Quartz cron expression fires after now, as a JSON string in \"EEE MMM dd HH:mm:ss zzz yyyy\" format.",
        "operationId" : "getNextCleanupRun",
        "parameters" : [
          {
            "description" : "Quartz cron expression, starting with the seconds field.",
            "in" : "query",
            "name" : "cronExpression",
            "required" : true,
            "schema" : {
              "type" : "string"
            }
          }
        ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "type" : "string"
                }
              }
            },
            "description" : "OK"
          }
        },
        "summary" : "Get the next run time of a schedule",
        "tags" : [
          "data-cleanup-controller"
        ]
      }
    },
    "/api/tdm/cleanup/run" : {
      "post" : {
        "description" : "Runs the cleanup configuration from the request body, without saving it, on the table in tableName and on the tables of the same system in the environments listed in environmentsList. Returns the number of checked and removed rows per table.",
        "operationId" : "runDataCleanup",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/CleanupSettings"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/CleanupResults"
                  }
                }
              }
            },
            "description" : "OK"
          }
        },
        "summary" : "Run a cleanup now",
        "tags" : [
          "data-cleanup-controller"
        ]
      }
    },
...
    "/api/tdm/rest/create-env" : {
      "delete" : {
        "description" : "Deletes the system in systemDeleteName, or the whole environment when systemDeleteName is not set. Returns HTTP 404 when the environment does not exist.",
        "operationId" : "deleteEnvironment",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/EnvironmentManagementRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ResponseMessage"
                }
              }
            },
            "description" : "OK"
          },
          "404" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ResponseMessage"
                }
              }
            },
            "description" : "Not Found"
          }
        },
        "summary" : "Delete an environment or a system",
        "tags" : [
          "atp-env-controller"
        ]
      },
      "post" : {
        "description" : "Creates the environment with the system and its connection, or adds the system to an existing environment. Returns HTTP 400 when the system already exists.",
        "operationId" : "createEnvironment",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/EnvironmentManagementRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ResponseMessage"
                }
              }
            },
            "description" : "OK"
          },
          "404" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ResponseMessage"
                }
              }
            },
            "description" : "Not Found"
          }
        },
        "summary" : "Create an environment or add a system",
        "tags" : [
          "atp-env-controller"
        ]
      },
      "put" : {
        "description" : "Replaces the connection of the system, and renames the environment to newEnvName or the system to newSystemName when they are set.",
        "operationId" : "updateEnvironment",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/EnvironmentManagementRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ResponseMessage"
                }
              }
            },
            "description" : "OK"
          },
          "404" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ResponseMessage"
                }
              }
            },
            "description" : "Not Found"
          }
        },
        "summary" : "Update a connection, or rename an environment or system",
        "tags" : [
          "atp-env-controller"
        ]
      }
    },
...
  },
  "components" : {
    "schemas" : {
...
      "CleanupResults" : {
        "type" : "object",
        "properties" : {
          "error" : {
            "type" : "string"
          },
          "recordsRemoved" : {
            "type" : "integer",
            "format" : "int32"
          },
          "recordsTotal" : {
            "type" : "integer",
            "format" : "int32"
          },
          "tableName" : {
            "type" : "string"
          }
        }
      },
      "CleanupSettings" : {
        "type" : "object",
        "properties" : {
          "environmentsList" : {
            "type" : "array",
            "items" : {
              "type" : "string",
              "format" : "uuid"
            }
          },
          "tableName" : {
            "type" : "string"
          },
          "testDataCleanupConfig" : {
            "$ref" : "#/components/schemas/TestDataCleanupConfig"
          }
        }
      },
...
      "EnvironmentConnectionRequest" : {
        "type" : "object",
        "properties" : {
          "name" : {
            "type" : "string"
          },
          "parameters" : {
            "type" : "object",
            "additionalProperties" : {
              "type" : "string"
            }
          },
          "type" : {
            "type" : "string"
          }
        }
      },
      "EnvironmentManagementRequest" : {
        "type" : "object",
        "properties" : {
          "connection" : {
            "$ref" : "#/components/schemas/EnvironmentConnectionRequest"
          },
          "envName" : {
            "type" : "string"
          },
          "environment" : {
            "type" : "object",
            "additionalProperties" : {
              "type" : "object"
            },
            "writeOnly" : true
          },
          "newEnvName" : {
            "type" : "string"
          },
          "newSystemName" : {
            "type" : "string"
          },
          "projectName" : {
            "type" : "string"
          },
          "systemDeleteName" : {
            "type" : "string"
          },
          "systemName" : {
            "type" : "string"
          }
        }
      },
...
      "ResponseMessage" : {
        "type" : "object",
        "properties" : {
          "content" : {
            "type" : "string"
          },
          "contentObject" : {
            "type" : "object"
          },
          "link" : {
            "type" : "string"
          },
          "type" : {
            "type" : "string",
            "enum" : [
              "success",
              "error"
            ]
          }
        }
      },
...
      "TestDataCleanupConfig" : {
        "type" : "object",
        "properties" : {
          "enabled" : {
            "type" : "boolean"
          },
          "id" : {
            "type" : "string",
            "format" : "uuid"
          },
          "queryTimeout" : {
            "type" : "integer",
            "format" : "int32"
          },
          "schedule" : {
            "type" : "string"
          },
          "scheduled" : {
            "type" : "string"
          },
          "searchClass" : {
            "type" : "string"
          },
          "searchDate" : {
            "type" : "string"
          },
          "searchSql" : {
            "type" : "string"
          },
          "shared" : {
            "type" : "boolean"
          },
          "type" : {
            "type" : "string",
            "enum" : [
              "SQL",
              "DATE",
              "CLASS"
            ]
          }
        }
      },
...
    },
    "securitySchemes" : {
      "Bearer Authentication" : {
        "bearerFormat" : "JWT",
        "scheme" : "bearer",
        "type" : "http"
      }
    }
  }
}